requirements.txt
PyGithub==1.55
cookiecutter==1.7.3